+++ src/Controller/StagiaireController.php
@@ -17,6 +17,7 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 class StagiaireController extends AbstractController
 {
     #[Route('/stagiaire', name: 'app_stagiaire')]
+    #[IsGranted('IS_AUTHENTICATED_FULLY')]
     public function index(EntityManagerInterface $entityManager): Response
     {
         //SELECT * FROM  stagiaire ORDER BY name ASC
@@ -31,37 +32,43 @@ class StagiaireController extends AbstractController
     #[IsGranted('IS_AUTHENTICATED_FULLY')]
     public function new(Stagiaire $stagiaire = null, Request $request, EntityManagerInterface $entityManager): Response
     {
-        if(!$stagiaire){
-            $stagiaire = new Stagiaire();
-        }
-        $form = $this->createForm(StagiaireType::class, $stagiaire);
-
-        $form->handleRequest($request);
-
-        if($form->isSubmitted() && $form->isValid()) {
-
-            $stagiaire = $form->getData();
-            //prepare PDO
-            $entityManager->persist($stagiaire);
-            //execute PDO
-            $entityManager->flush();
-
+        if (!$this->isGranted("ROLE_ADMIN")) {
             return $this->redirectToRoute('app_stagiaire');
-
+        } else {
+            if(!$stagiaire){
+                $stagiaire = new Stagiaire();
+            }
+            $form = $this->createForm(StagiaireType::class, $stagiaire);
+    
+            $form->handleRequest($request);
+    
+            if($form->isSubmitted() && $form->isValid()) {
+    
+                $stagiaire = $form->getData();
+                //prepare PDO
+                $entityManager->persist($stagiaire);
+                //execute PDO
+                $entityManager->flush();
+    
+                return $this->redirectToRoute('app_stagiaire');
+    
+            }
+    
+            return $this->render('stagiaire/new.html.twig', [
+                'formAddStagiaire' => $form,
+                'edit' => $stagiaire->getId()
+            ]);
         }
-
-        return $this->render('stagiaire/new.html.twig', [
-            'formAddStagiaire' => $form,
-            'edit' => $stagiaire->getId()
-        ]);
     }
 
     #[Route('/stagiaire/{id}/remove', name: 'remove_stagiaire')]
     #[IsGranted('IS_AUTHENTICATED_FULLY')]
     public function remove(Stagiaire $stagiaire, EntityManagerInterface $entityManager): Response
     {
-        $entityManager->remove($stagiaire);
-        $entityManager->flush();
+        if($this->isGranted("ROLE_ADMIN")){
+            $entityManager->remove($stagiaire);
+            $entityManager->flush();
+        }
         
         return $this->redirectToRoute('app_stagiaire');
     }
@@ -72,26 +79,28 @@ class StagiaireController extends AbstractController
     #[ParamConverter("stagiaire", options: ['mapping' => ["idStagiaire" => "id"]])]
     #[ParamConverter("session", options: ['mapping' => ["idSession" => "id"]])]
     public function add_remove_session(Session $session, Stagiaire $stagiaire, EntityManagerInterface $entityManager) {
-        
-        $stagiaireSubscribed = $entityManager->getRepository(Stagiaire::class)->findStagiaireArrayInSessionId($session->getId());
+        if($this->isGranted("ROLE_ADMIN")) {
+            $stagiaireSubscribed = $entityManager->getRepository(Stagiaire::class)->findStagiaireArrayInSessionId($session->getId());
 
-        if(in_array($stagiaire, $stagiaireSubscribed)){
-            $stagiaire->removeSession($session);
+            if(in_array($stagiaire, $stagiaireSubscribed)){
+                $stagiaire->removeSession($session);
+            }
+            else{
+                $stagiaire->addSession($session);
+            }
+
+            $entityManager->persist($session);
+            $entityManager->persist($stagiaire);
+
+            $entityManager->flush();
         }
-        else{
-            $stagiaire->addSession($session);
-        }
-
-        $entityManager->persist($session);
-        $entityManager->persist($stagiaire);
-
-        $entityManager->flush();
 
         return $this->redirectToRoute('show_stagiaire', ['id' => $stagiaire->getId()]);
     } 
 
     
     #[Route('/stagiaire/{id}', name: 'show_stagiaire')]
+    #[IsGranted('IS_AUTHENTICATED_FULLY')]
     public function show(EntityManagerInterface $entityManager,Stagiaire $stagiaire): Response
     {
         if($stagiaire){
